+++ CS665/main/io/ReadFile.java
@@ -6,13 +6,13 @@ import java.util.Vector;
 
 public interface ReadFile {
 	
-	public Vector<String> readUserFile(String givenUserName, int givenUserID, String givenUserType);
+	public Vector<String> readUserFile(String givenUserName, long givenUserID, String givenUserType);
 
 	
-	public Vector<String> readRosterFile(String userName, int userID);
+	public Vector<String> readRosterFile(String userName, long userID);
 
 
-	public Vector<String> readInventoryFile(String givenInvName, int givenInvID) throws IOException, FileNotFoundException;
+	public Vector<String> readInventoryFile(String givenInvName, long givenInvID) throws IOException, FileNotFoundException;
 	
 	
 }//end ReadFile
